Move initSidebarReferenceFont to ReferenceFontPanel.init
